🚚 api: move controller to results context
