+++ api/src/certification/results/application/certification-results-controller.js
@@ -1,7 +1,9 @@
 import dayjs from 'dayjs';
 
+import { tokenService } from '../../../shared/domain/services/token-service.js';
 import { usecases } from '../domain/usecases/index.js';
 import { getCleaCertifiedCandidateCsv } from '../infrastructure/utils/csv/certification-results/get-clea-certified-candidate-csv.js';
+import { getSessionCertificationResultsCsv } from '../infrastructure/utils/csv/certification-results/get-session-certification-results-csv.js';
 
 const getCleaCertifiedCandidateDataCsv = async function (request, h, dependencies = { getCleaCertifiedCandidateCsv }) {
   const sessionId = request.params.id;
@@ -21,6 +23,31 @@ const getCleaCertifiedCandidateDataCsv = async function (request, h, dependencie
     .header('Content-Disposition', `attachment; filename=${fileName}`);
 };
 
-const certificationResultsController = { getCleaCertifiedCandidateDataCsv };
+const getSessionResultsByRecipientEmail = async function (
+  request,
+  h,
+  dependencies = { tokenService, getSessionCertificationResultsCsv },
+) {
+  const token = request.params.token;
+
+  const { resultRecipientEmail, sessionId } =
+    dependencies.tokenService.extractCertificationResultsByRecipientEmailLink(token);
+  const { session, certificationResults } = await usecases.getSessionResultsByResultRecipientEmail({
+    sessionId,
+    resultRecipientEmail,
+  });
+  const csvResult = await dependencies.getSessionCertificationResultsCsv({
+    session,
+    certificationResults,
+    i18n: request.i18n,
+  });
+
+  return h
+    .response(csvResult.content)
+    .header('Content-Type', 'text/csv;charset=utf-8')
+    .header('Content-Disposition', `attachment; filename=${csvResult.filename}`);
+};
+
+const certificationResultsController = { getCleaCertifiedCandidateDataCsv, getSessionResultsByRecipientEmail };
 
 export { certificationResultsController };
