+++ api/tests/certification/results/unit/application/certification-results-controller_test.js
@@ -1,8 +1,9 @@
 import { certificationResultsController } from '../../../../../src/certification/results/application/certification-results-controller.js';
 import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
+import { getI18n } from '../../../../tooling/i18n/i18n.js';
 
-describe('Certification | Results | Unit | Controller | session-with-clea-certified-candidate', function () {
+describe('Certification | Results | Unit | Controller | certification results', function () {
   describe('#getCleaCertifiedCandidateDataCsv', function () {
     it('should return a response with CSV results', async function () {
       // given
@@ -42,6 +43,48 @@ describe('Certification | Results | Unit | Controller | session-with-clea-certif
       expect(response.headers['Content-Type']).to.equal('text/csv;charset=utf-8');
       expect(response.headers['Content-Disposition']).to.equal(
         'attachment; filename=20210101_1430_candidats_certifies_clea_1.csv',
+      );
+    });
+  });
+
+  describe('#getSessionResultsByRecipientEmail ', function () {
+    it('should return csv content and fileName', async function () {
+      // given
+      const i18n = getI18n();
+      const session = { id: 1, date: '2020/01/01', time: '12:00' };
+      const dependencies = {
+        getSessionCertificationResultsCsv: sinon.stub(),
+        tokenService: {
+          extractCertificationResultsByRecipientEmailLink: sinon.stub(),
+        },
+      };
+      dependencies.tokenService.extractCertificationResultsByRecipientEmailLink
+        .withArgs('abcd1234')
+        .returns({ sessionId: 1, resultRecipientEmail: 'user@example.net' });
+
+      sinon
+        .stub(usecases, 'getSessionResultsByResultRecipientEmail')
+        .withArgs({ sessionId: session.id, resultRecipientEmail: 'user@example.net' })
+        .resolves({
+          session,
+          certificationResults: [],
+        });
+
+      dependencies.getSessionCertificationResultsCsv
+        .withArgs({ session, certificationResults: [], i18n })
+        .resolves({ content: 'csv content', filename: '20200101_1200_resultats_session_1.csv' });
+
+      // when
+      const response = await certificationResultsController.getSessionResultsByRecipientEmail(
+        { i18n, params: { token: 'abcd1234' } },
+        hFake,
+        dependencies,
+      );
+
+      // then
+      expect(response.source).to.deep.equal('csv content');
+      expect(response.headers['Content-Disposition']).to.equal(
+        'attachment; filename=20200101_1200_resultats_session_1.csv',
       );
     });
   });
