+++ api/tests/unit/application/session/session-controller_test.js
@@ -54,48 +54,6 @@ describe('Unit | Controller | sessionController', function () {
     });
   });
 
-  describe('#getSessionResultsByRecipientEmail ', function () {
-    it('should return csv content and fileName', async function () {
-      // given
-      const i18n = getI18n();
-      const session = { id: 1, date: '2020/01/01', time: '12:00' };
-      const dependencies = {
-        getSessionCertificationResultsCsv: sinon.stub(),
-        tokenService: {
-          extractCertificationResultsByRecipientEmailLink: sinon.stub(),
-        },
-      };
-      dependencies.tokenService.extractCertificationResultsByRecipientEmailLink
-        .withArgs('abcd1234')
-        .returns({ sessionId: 1, resultRecipientEmail: 'user@example.net' });
-
-      sinon
-        .stub(usecases, 'getSessionResultsByResultRecipientEmail')
-        .withArgs({ sessionId: session.id, resultRecipientEmail: 'user@example.net' })
-        .resolves({
-          session,
-          certificationResults: [],
-        });
-
-      dependencies.getSessionCertificationResultsCsv
-        .withArgs({ session, certificationResults: [], i18n })
-        .resolves({ content: 'csv content', filename: '20200101_1200_resultats_session_1.csv' });
-
-      // when
-      const response = await sessionController.getSessionResultsByRecipientEmail(
-        { i18n, params: { token: 'abcd1234' } },
-        hFake,
-        dependencies,
-      );
-
-      // then
-      expect(response.source).to.deep.equal('csv content');
-      expect(response.headers['Content-Disposition']).to.equal(
-        'attachment; filename=20200101_1200_resultats_session_1.csv',
-      );
-    });
-  });
-
   describe('#getSessionResultsToDownload ', function () {
     it('should return results to download', async function () {
       // given
